pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "iamspy"
version = "0.1.0"
description = "Spy on your Python code to find exactly which GCP IAM permissions it needs"
requires-python = ">=3.12"
dependencies = [
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "aiofiles>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "vulture>=2.0",
]
build = [
    "anthropic>=0.30",
    "google-genai>=1.0",
    "google-cloud-iam>=2.0",
    "sentence-transformers>=3.0",
    "einops>=0.8",
]

[project.scripts]
iamspy = "iamspy.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests that require network, installed GCP SDKs, or performance benchmarks",
]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests", "build_pipeline"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "C90", "PLR"]
ignore = ["E501", "PLR2004"]

[tool.ruff.lint.pylint]
max-statements = 50
max-branches = 12
max-args = 8

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.isort]
known-first-party = ["iamspy", "build_pipeline"]

[tool.coverage.run]
source = ["iamspy"]

[tool.coverage.report]
show_missing = true
skip_empty = true
